requirements.txt
cryptography==44.0.0
infi.systray==0.1.12
pywebview==5.3.2
requests==2.32.3
schedule==1.2.2
